Multiple bug fixes, parser changes, and error checking.  Increased error checking in the assertion storage class with increased checking of file writing.  Bug fixes with the assertion storage and parser, including formatting changes to the saved assertion stores - all .json now, and formatting changes to the parser, requiring .json format for assertion store parsing.  In addition, parser now includes abilities to parse assertion store array property class values.

+++ SenTest/SenTest/Sentegrity/TrustScore Computation/Sentegrity_TrustScore_Computation.m
@@ -200,7 +200,7 @@ NSMutableArray *triggeredTrustFactorOutputObjects;
                 
                 // Set the penalty weight for the subclass
                 subClass.weightedPenalty = (subClass.basePenalty * (1-(0.1 * subClass.weight.integerValue)) );
-                NSLog(@"Subclass Name:%@ %d",subClass.name,subClass.weightedPenalty);
+                NSLog(@"Subclass Name:%@ %ld",subClass.name,(long)subClass.weightedPenalty);
                 
                 // Add the subclass weightedPenalty to the classification basePenalty
                 class.basePenalty = (class.basePenalty + subClass.weightedPenalty);
